If block

+++ install.sh
@@ -1,7 +1,9 @@
 #!/bin/sh
 set -e
 
-cp ~/.vimrc ~/.vimrc.pre
+if [ -f ~/.vimrc ]; then
+  cp ~/.vimrc ~/.vimrc.pre
+fi
 
 cp ~/.vim_conf/.vimrc ~/.vimrc
 echo "Installed successfully"
